feat(FAT1-67):edit some comments
