+++ lib/presentation/profile/reset_password/views/widgets/profile_reset_password_app_bar.dart
@@ -5,9 +5,9 @@ import 'package:flowery_tracking_app/utils/common_widgets/custom_back_arrow.dart
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 
-class ResetPasswordAppBar extends StatelessWidget
+class ProfileResetPasswordAppBar extends StatelessWidget
     implements PreferredSizeWidget {
-  const ResetPasswordAppBar({super.key});
+  const ProfileResetPasswordAppBar({super.key});
 
   @override
   Widget build(BuildContext context) {
@@ -16,7 +16,10 @@ class ResetPasswordAppBar extends StatelessWidget
       titleWidget: RPadding(
         padding: const EdgeInsets.symmetric(horizontal: 16),
         child: Row(
-          children: [const CustomBackArrow(), Text(AppText.password.tr())],
+          children: [
+            const CustomBackArrow(),
+            Text(AppText.resetPasswordTitle.tr()),
+          ],
         ),
       ),
     );
