+++ lib/presentation/profile/reset_password/views_model/profile_reset_password_intent.dart
@@ -6,7 +6,7 @@ class InitializeProfileResetPasswordFormIntent
     extends ProfileResetPasswordIntent {}
 
 class OnProfileResetPasswordIntent extends ProfileResetPasswordIntent {
-  ProfileResetPasswordRequestEntity request;
+  final ProfileResetPasswordRequestEntity request;
 
   OnProfileResetPasswordIntent({required this.request});
 }
